research_agent/requirements.txt
# Research Agent Dependencies
openai>=1.99.3
notion-client==2.0.0  # Pinned to 2.0.0 - v2.7.0+ removed databases.query in favor of data_sources.query (not yet fully deployed)
requests>=2.31.0
python-dotenv>=1.0.0
pytz>=2024.1
boto3>=1.26.0
botocore>=1.29.0